Fully destroy endpoints when refcount is 0
Previous to this patch, endpoints were never freed, but just had most
resources released when the ref count reached zero.  This was because
the use of pthread thread specific data store meant that only the
thread that created the ep could remove it from the thread specific
data store.  Since we couldn't guarantee that, we never actually
removed data from the store.

This patch switches from pthread thread specific storage to using a
hash table indexed by the thread id.  Note that we use gettid() instead
of pthread_self(), because pthread_t is a opaque structure that is not
comparable and so can't really be used as a hash key.  If we want to
support OSes other than Linux, then we can add more ways to get the
unique thread id.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>