fix: sendrecv/listen: silence leak warning.
gcc14 incorrectly flagged this as a leak as it was unable to reason
about the way that we are storing into a listen_comm_t, but have
allocated a sendrecv_listen_comm_t.

This function can be rewritten to be much more clear, though, and doing
so silences the warning. Cast to the base struct when assigning into the
inout ptr, Stop calling fi_strerror(-ret) as ret was never set to any
error code. gotos are unnecessary in this function, there is no valid
path where we allocate memory and do not return 0, so just return errors
directly.

Signed-off-by: Nicholas Sielicki <nslick@amazon.com>